requirements.txt
Flask
requests
python-dotenv
SpeechRecognition
pydub
gtts